debug bbox Rect2i to 2f

+++ ch6_optical_flow/src/bounding_box.h
@@ -25,5 +25,5 @@ class BoundingBox {
     }
 
    private:
-    cv::Rect2i window_;
+    cv::Rect2f window_;
 };
